requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
backoff==2.2.1
bcrypt==4.1.3
blinker==1.9.0
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.31.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==30.1.0
langchain==0.3.17
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langsmith==0.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
narwhals==1.24.2
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.61.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
posthog==3.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.25.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.0
pydantic_core==2.20.0
pydeck==0.9.1
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.37.2
streamlit==1.41.1
sympy==1.12.1
tenacity==8.4.2
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.18.3
zipp==3.19.2
zstandard==0.23.0
